requirements.txt
crewai

langgraph
langchain
langchain-ollama

autogen-agentchat

autogen-agentchat
autogen-ext[openai]

python-dotenv